django/requirements.txt
django>=5.0,<6.0
psycopg2-binary
pillow
django-jinja
jinja2-simple-tags
django-extensions
pytz
requests
beautifulsoup4
djangorestframework
markdown
django-filter
jieba
lxml[html_clean]
smartypants
num2words
chinese-converter
openai
celery
redis
channels[daphne]
channels_redis
xxhash
json5
pydantic
inflect
tabulate
markupsafe==2.1.5

ipython
django-livereload-server
django-debug-toolbar


